requirements.txt
google-generativeai>=0.3.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
yt-dlp>=2024.1.0
openai-whisper>=20231117
Pillow>=10.0.0
opencv-python-headless>=4.8.0
requests>=2.31.0
acrcloud>=1.0.0
